サイドバーのコンポーネント分割
